refactor rats_e2e.aml
