+++ rats-devtools/src/rats_e2e/aml/__init__.py
@@ -1,9 +1,10 @@
 """Examples and end-to-end tests for the `rats.aml` package."""
 
-from ._app import Application
-from ._example_cli import ExampleCliApp
+from ._app import Application, main
+from ._example_job import ExampleJob
 
 __all__ = [
     "Application",
-    "ExampleCliApp",
+    "ExampleJob",
+    "main",
 ]
