+++ rats-devtools/src/rats_e2e/aml/__main__.py
@@ -6,11 +6,10 @@ Run this module with coverage.py: `coverage run --branch -m rats_e2e.runtime`.
 
 import logging
 
-from rats import apps
 from rats_e2e import aml
 
 logger = logging.getLogger(__name__)
 
 
 if __name__ == "__main__":
-    apps.run_plugin(aml.ExampleCliApp)
+    aml.main()
